all get_pipeline_ functions divided by task
