+++ examples/advanced/time_series_forecasting/composing_pipelines.py
@@ -18,7 +18,8 @@ from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum, TsForecastingParams
-from examples.simple.time_series_forecasting.pipelines import *
+from examples.simple.time_series_forecasting.ts_pipelines import *
+from fedot.core.utils import fedot_project_root
 
 
 def get_available_operations():
@@ -30,10 +31,10 @@ def get_available_operations():
 
 
 datasets = {
-    'australia': '../data/ts/australia.csv',
-    'beer': '../data/ts/beer.csv',
-    'salaries': '../data/ts/salaries.csv',
-    'stackoverflow': '../data/ts/stackoverflow.csv'}
+    'australia': f'{fedot_project_root()}/examples/data/ts/australia.csv',
+    'beer': f'{fedot_project_root()}/examples/data/ts/beer.csv',
+    'salaries': f'{fedot_project_root()}/examples/data/ts/salaries.csv',
+    'stackoverflow': f'{fedot_project_root()}/examples/data/ts/stackoverflow.csv'}
 
 
 def run_composing(dataset: str, pipeline: Pipeline, len_forecast=250):
