+++ examples/advanced/time_series_forecasting/exogenous.py
