+++ examples/advanced/time_series_forecasting/multistep.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 
 from examples.advanced.time_series_forecasting.composing_pipelines import get_border_line_info
-from examples.simple.time_series_forecasting.pipelines import *
+from examples.simple.time_series_forecasting.ts_pipelines import *
 from examples.simple.time_series_forecasting.tuning_pipelines import visualise
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
@@ -13,14 +13,17 @@ from fedot.core.pipelines.ts_wrappers import out_of_sample_ts_forecast
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import TaskTypesEnum, Task, TsForecastingParams
 
+from fedot.core.utils import fedot_project_root
+
 warnings.filterwarnings('ignore')
 np.random.seed(2020)
 
 datasets = {
-    'australia': '../data/ts/australia.csv',
-    'beer': '../data/ts/beer.csv',
-    'salaries': '../data/ts/salaries.csv',
-    'stackoverflow': '../data/ts/stackoverflow.csv'}
+    'australia': f'{fedot_project_root()}/examples/data/ts/australia.csv',
+    'beer': f'{fedot_project_root()}/examples/data/ts/beer.csv',
+    'salaries': f'{fedot_project_root()}/examples/data/ts/salaries.csv',
+    'stackoverflow': f'{fedot_project_root()}/examples/data/ts/stackoverflow.csv'}
+
 
 
 def run_multistep(dataset: str, pipeline: Pipeline, step_forecast=10):
