+++ test/unit/pipelines/test_decompose_pipelines.py
@@ -2,7 +2,7 @@ from itertools import product
 
 import numpy as np
 
-from examples.advanced.classification_with_tuning_example import get_classification_dataset
+from examples.simple.classification.classification_with_tuning import get_classification_dataset
 from examples.advanced.decompose.refinement_forecast_example import get_refinement_pipeline
 from fedot.core.data.data import InputData
 from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
